certbotsh: Strictly check for errors while installing new data
It still does not solve nesting issues observed on deployments where
contents of $t ($domain) not owned by $runas user (e.g. default
letsencrypt user), so rm -rf "$td" in exit_handler() cannot remove it
causing directory nesting on each run, but make sure errors catched at
each stage of new data installation process.

+++ certbot.sh
@@ -1833,11 +1833,13 @@ applet()
         done
 
         # Install new (temporary) data
-        mv -f "$t" "$td" 2>/dev/null ||:
-        if mv -f "$td" "$t"; then
-            rm -rf "$t/$domain" && chmod 0755 "$t" && td='' ||:
-        else
-            mv -f "$td/$domain" "$t" 2>/dev/null ||:
+        if mv -f "$t" "$td"; then
+            if mv -f "$td" "$t"; then
+                chmod 0755 "$t" ||:
+                td="$t/$domain"
+            else
+                mv -f "$td/$domain" "$t" ||:
+            fi
         fi
 
         ## Leave renewed certificate lineage
